show active dot, desktop: better hover ux
Before, when you hovered out of a child menu, the menu would close. This makes it too easy to trigger a close. Instead, make the menu work like the macOS context menu. In this menu, the child menu only closes when the parent chooses a different menu item.

In the future, it would also be good to delay open and closing of these menus